requirements.txt
# Core
langchain>=1.0.0
langchain-openai>=0.3.0
langchain-ollama>=0.3.0
langgraph>=0.4.0
deepagents>=0.4.0

# MCP
mcp>=1.0.0
langchain-mcp-adapters>=0.1.0

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sse-starlette>=2.0.0

# Data validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Storage
redis>=5.0.0
aioredis>=2.0.0

# Utils
pyyaml>=6.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
aiofiles>=24.0.0

# Sandbox (Docker SDK for run_in_sandbox)
docker>=7.0.0

# Logging
loguru>=0.7.0

# Task mode: schedule trigger (cron)
apscheduler>=3.10.0
